Switch to uint16 to allow for larger number of chunks

+++ include/render_system/fog/sync/signal/CalcFromFt.hpp
@@ -8,12 +8,12 @@ namespace render_system::fog::sync::signal
 {
 class CalcFromFt
 {
-    uint8_t m_myChunkOrder{0};
-    uint8_t m_totalNumChunks{0};
+    uint16_t m_myChunkOrder{0};
+    uint16_t m_totalNumChunks{0};
     const star::common::FrameTracker *m_ft{nullptr};
 
   public:
-    CalcFromFt(uint8_t myChunkOrder, uint8_t totalNumChunks, const star::common::FrameTracker *ft)
+    CalcFromFt(uint16_t myChunkOrder, uint16_t totalNumChunks, const star::common::FrameTracker *ft)
         : m_myChunkOrder(std::move(myChunkOrder)), m_totalNumChunks(std::move(totalNumChunks)), m_ft(ft)
     {
     }
