+++ include/render_system/fog/sync/wait/WaitForFirstChunk.hpp
@@ -8,11 +8,11 @@ namespace render_system::fog::sync::wait
 {
 class WaitForFirstChunk
 {
-    uint8_t m_totalNumChunks{0};
+    uint16_t m_totalNumChunks{0};
     const star::common::FrameTracker *m_ft{nullptr};
 
   public:
-    WaitForFirstChunk(uint8_t totalNumChunks, const star::common::FrameTracker *ft)
+    WaitForFirstChunk(uint16_t totalNumChunks, const star::common::FrameTracker *ft)
         : m_totalNumChunks{std::move(totalNumChunks)} ,m_ft{ft}
     {
     }
